feat: show last forged block for active round (#294)

+++ app/Http/Livewire/MonitorStatistics.php
@@ -29,7 +29,7 @@ final class MonitorStatistics extends Component
 
     private function delegateRegistrations(): int
     {
-        return Cache::remember('delegateRegistrations', Network::blockTime(), function (): int {
+        return (int) Cache::remember('delegateRegistrations', Network::blockTime(), function (): int {
             return Transaction::withScope(DelegateRegistrationScope::class)->count();
         });
     }
